Allow cache-control to be configured by system property
The Cache-Control request header determines whether or not requests
made by the ECF HTTP transfers use cached values from intermediate
proxies or not.

Due to bug 249990 caching was intentionally disabled. However this
causes problems when proxies are configured correctly from taking
advantage of caching.

Exposing this as a system property allows the age to be controlled
externally from ECF, and can be completely disabled by setting

 -Dorg.eclipse.ecf.http.cache.max-age=-1

Alternatively, a non-zero number represents a maximum age that
can be requested, and is passed through to the server.

Bug: 410813
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>
Change-Id: If231f3f5d88a146b38e56836bf69d5c25b162fd1