Bug 396457 - [Releng][Maven] Create a build based on tycho
First part of adding the necessary machinery to build ECF with tycho.
Adds two new projects, one to contain target platform definitions,
and one to contain p2 repository definitions. Adds a parent pom at
the root of the project and a gitignore file to mask built artifacts
from the git repo. Poms are added and the build enabled for all the
core features and plug-ins used by Equinox P2 (i.e. the minimum
necessary to build the Eclipse Platform itself.)

Change-Id: I0b41b4a611e2b3588fe876f14bd7c8e80de07b67
Signed-off-by: Mat Booth <mat.booth@redhat.com>