Bug 495684 - [ui] o.e.ecf.remoteservice.ui does not build with tycho
The package import on "org.eclipse.ui.views.properties" is a "split
package" (i.e. this package is provided by more than one bundle in
the target platform and the requirement here is ambiguous) so
Equinox is not able to accurately wire the correct dependency.

Adding the extra parameter: ;ui.views="split"

Helps Equinox to find and wire the correct bundle dependency.

Change-Id: I6cc1336a85c6aec992baf11d18f40ef9378cf4c1
Signed-off-by: Mat Booth <mat.booth@redhat.com>