Bug 495773 - [sync] o.e.ecf.sync does not build with tycho
The package import on "org.eclipse.jface.text" and "org.eclipse.
core.runtime" are "split packages" (i.e. the packages are provided
by more than one bundle in the target platform and the requirement
is ambiguous) so Equinox is not able to accurately wire the correct
dependency.

Adding the extra parameters helps Equinox to find and wire the
correct bundle dependency.

Change-Id: Ib6abc7c2f05b586871b389eade21bfff9b802db0
Signed-off-by: Mat Booth <mat.booth@redhat.com>