Bug 508039 - [releng] build "com.mycorp.examples.timeservice.async"
Add the bundle to the build so that karaf maven installations work
correctly. Also add the bundle to the examples feature (so it gets
published in the p2 repo) so that karaf http installations work
correctly.

Fix also the incorrect bundle version in the manifest, since
previously this bundle was published with a version of 2.0.0.

Change-Id: Ideb322b3ba52935e307c317a8a1bc311cd5fb669
Signed-off-by: Mat Booth <mat.booth@redhat.com>