Fix for bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=534087
Allow ECF to build against both R6 and R7 implementations of the
org.osgi.service.log API. This is accomplished by providing stubs
of the new interfaces on the build-time classpath so that we can
implement the new R7 methods without having unresolved classes on
R6.

The stubs are not shipped with the binary distribution of ECF,
so in the actual ECF code we can use them as long as we avoid any
explicit import statements to avoid runtime linkage errors on R6.

Change-Id: Ib29135c030e3dd8a3bf799552a2f3144c26edfe3
Signed-off-by: Mat Booth <mat.booth@redhat.com>