Bug 544447 - [provider] Implement filetransfer provider using HttpClient 4.5 API
The guard in HttpClientRetrieveFileTransfer.hardClose() is wrong. I
imagine the httpResponse should be closed when we are done but not
paused rather than if we are not done but paused which sounds like the
opposite because we should certainly not close when we are paused!

Anyway, someone can confirm what the logic should be, but certainly like
this the SetupArchiver works properly again and leases are released
properly, i.e., I see mostly just pairs of the operations and the
per-route count never exceeds the number of threads that I'm using to
load in parallel.

Change-Id: I4285b278278c9f3cb7c6d1bae4d8e8b335e90b37
Signed-off-by: Ed Merks <ed.merks@gmail.com>