Bug 544447 - [provider] Implement filetransfer provider using HttpClient 4.5 API
Make sure connection is closed in hardClose().

There are four cases to consider:

1. Transfer is complete
2. Transfer was cancelled
3. There was an exception during transfer
4. The transfer was paused

In case 3, hardClose() is called from cancel() directly, in the other
cases it gets called from the finally block in performFileTransfer().

Cases 1-3 are easy: In case 1, the stream encountered EOF, isDone() ==
true, and the connection was already closed as a result. In case 2
and 3, it is obvious that we want the connection closed as a result.

In case 4, we might want to resume later, but when resuming, we are
actually creating a new HttpGet request and get a new HttpResponse as a
result. So we also want to close the previous one when pausing.

Change-Id: I3daf3c34d0784e5c19aa38441ba373422796f538
Signed-off-by: Carsten Reckord <reckord@yatta.de>