Bug 560808 - The download.eclipse.org server appears to be dropping connections
This change adds a new configuration option for the httpclient45
provider to set the TTL for connections. It now defaults to 1000ms
instead of -1 (infinite) to help improve performance when talking to
servers that use very short TTLs such as dl.e.org.

This is for the httpclient45 provider only since old versions of
Apache HTTP Components that may be used by the old httpclient4
provider does not support setting the TTL.

Change-Id: I768d9e67e48d4c1f3afe00036276befb421dbc89
Signed-off-by: Mat Booth <mat.booth@redhat.com>