Bug 566100 - Remove explicit dependency on JNA
Win32HttpClientConfigurationModifier.isWinAuthAvailable() does nothing
useful. It would only return false if com.sun.jna.platform.win32.Sspi
would not be available.

But this can never be the case, since we depend on
org.apache.httpcomponents.httpclient.win, which in turn depends on JNA
(non-optional).

The original code in
org.apache.http.impl.client.WinHttpClients.isWinAuthAvailable() is a
no-op for the same reason.

Change-Id: Ic88ec9770dbc5bc296701791348dbf5a0c3a8a8d
Signed-off-by: Sebastian Ratz <sebastian.ratz@sap.com>