Add an experimental Certificate Chain sorter
Adds a simple experimental (disable by default) implementation to sort
a list of certificates and get a certificate chain. Some servers send
the certificates unordered. Until TLSv1.2 this was not allowed. TLSv1.3
allows it (see RFC 8446 section 4.4.2).

This is not at all fully complaint with the RFC 8446 as it does not
(currently) handles some cases like multiple chain paths or certificates
with the same CN (Common Name) that are in a specific order.