Allow http proxy authentication with SPNEGO on Windows
Nowadays companies often use proxy servers for malware detection.
Sometimes these servers need authentication (HTTP status code 407 Proxy
Authentication Required).

The direct use of NTLM authentication is now commonly replaced with
SPNEGO (RFC 4559) to negotiate an authentication scheme (e.g. Kerberos
or NTLM). In order to handle the response header "Proxy-Authenticate:
Negotiate", this scheme has to be added to the list of preferred
authentication schemes including NTLM.

In addition, the Win32HttpClientConfigurationModifier has to be
registered as a service implementation for IHttpClientModifier in order
to be found in Activator.getModifierReferences().